requirements.txt
flask==2.3.3
flask-cors==5.0.0
flask-wtf==1.1.1
firebase-admin==6.2.0
Pillow==10.0.0
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
gevent==24.2.1
rembg==2.0.55

# Optional dependencies for advanced features
# Uncomment if you want full AI capabilities
# torch>=2.2.0
# torchvision>=0.17.0
# opencv-python==4.8.0.74
# numpy==1.26.3 